重构: 从 cli_io 解耦出独立日志库 log_output
将原 cli_io.c 中高度耦合的日志打印核心（cli_printk/sys_printk）提取为
独立的 log_output 库，新增接口清晰的分层架构：

log_output.{h,c}
  - 零 CLI 依赖，可独立编译和移植
  - 内核风格级别过滤（KERN_EMERG..KERN_DEBUG）
  - ANSI 着色前缀
  - 可选终端 save/restore 钩子（term_ops）
  - 可选 kfifo 环形缓冲模式（LOG_OUTPUT_HAVE_KFIFO）
  - 独立 / 集成双模式（log_output_set_term_coord 切换）

cli_term.c
  - 新增终端适配层，桥接 log_output 与 LinCLI 交互状态
  - 导出 cli_term_save/restore，在 sys_printk 的 CS 外围
    做完整重绘（含候选列表），彻底解决 PC spinlock 死锁

cli_io.c
  - 移除 150+ 行日志实现，缩减至 340 行
  - pr_* / KERN_* / COLOR_* 宏保持向后兼容

cli_var.c
  - 修复 %u→%zu 格式警告（GCC __printf__ 属性检查）

+++ src/cli/cli_var.c
@@ -383,8 +383,7 @@ static int builtin_string_from_str(void *addr, size_t size, const char *str)
 		return -1;
 	size_t len = strlen(str);
 	if (len >= size) {
-		pr_warn("str trunc %u>%u\r\n", (unsigned int)len,
-			size - 1);
+		pr_warn("str trunc %zu>%zu\r\n", len, size - 1);
 		len = size - 1;
 	}
 	memcpy(addr, str, len);
